% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_background_correction.R
\name{correct_background}
\alias{correct_background}
\title{Background-correct and export results}
\usage{
correct_background(
  paths,
  FCS_export,
  CSV_export,
  chans = readRDS(file.path(paths["rds"], "chans.Rds")),
  transforms_chan = readRDS(file.path(paths["rds"], "transforms_chan.Rds")),
  transforms_pe = readRDS(file.path(paths["rds"], "transforms_pe.Rds")),
  xp = readRDS(file.path(paths["rds"], "xp.Rds")),
  xp_scaled = readRDS(file.path(paths["rds"], "xp_transformed_scaled.Rds")),
  umap = readRDS(file.path(paths["rds"], "umap.Rds")),
  events.code = readRDS(file.path(paths["rds"], "pe.Rds")),
  sampling = readRDS(file.path(paths["rds"], "sampling_preds.Rds")),
  preds = readRDS(file.path(paths["rds"], "predictions.Rds")),
  prediction_colnames = readRDS(file.path(paths["rds"], "prediction_colnames.Rds")),
  a = read.csv(paths["annotation"], sep = ",", header = TRUE, stringsAsFactors = FALSE),
  verbose = TRUE
)
}
\arguments{
\item{paths}{Character vector of paths to store intput, intermediary results, outputs...}

\item{FCS_export}{if "none", no FCS export. if "concatenated", export one FCS file with all the data. if "split", export the data in the result folder under the subfolder FCS, with each file corresponding to a (subsampled) input file.}

\item{CSV_export}{Boolean, whether to export results to CSV}

\item{chans}{vector of backbone channels' names}

\item{transforms_chan}{named list of logicle-transformations for backbone channels}

\item{transforms_pe}{named list of logicle-transformations for Infinity channels}

\item{xp}{Logicle-transformed backbone expression matrix}

\item{xp_scaled}{Per-well Z-scored Backbone expression matrix}

\item{umap}{UMAP dimensionality reduction matrix of the Backbone data}

\item{events.code}{vector of length nrow(xp) specifying from which well each event originates}

\item{sampling}{boolean vector of length nrow(xp) specifying which events were selected for final imputation}

\item{preds}{matrix of imputed data}

\item{prediction_colnames}{vector of column names for the imputed data, a subset of the column names in the preds matrix}

\item{a}{annotation table generated by infinityFlow:::intialize()}

\item{verbose}{Verbosity}
}
\description{
Background-correct and export results
}
