% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_svm_regression_all.R
\name{fit_svms}
\alias{fit_svms}
\title{Train SVM regressions}
\usage{
fit_svms(yvar, params = list(type = "eps-regression", cost = 1, epsilon =
  0.5), paths, cores, xp = readRDS(file.path(paths["rds"],
  "xp_transformed_scaled.Rds")), chans = readRDS(file.path(paths["rds"],
  "chans.Rds")), events.code = readRDS(file.path(paths["rds"],
  "pe.Rds")), transforms_chan = readRDS(file.path(paths["rds"],
  "transforms_chan.Rds")),
  transforms_pe = readRDS(file.path(paths["rds"], "transforms_pe.Rds")))
}
\arguments{
\item{yvar}{name of the exploratory measurement}

\item{params}{Named list of arguments passed to e1071:svm()}

\item{paths}{Character vector of paths to store intput, intermediary results, outputs...}

\item{cores}{Number of cores to use for parallel computing}
}
\description{
Train SVM regressions
}
