% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_regression_all.R
\name{fitter_nn}
\alias{fitter_nn}
\title{Wrapper to Neural Network training. Defined separetely to avoid passing too many objects in parLapplyLB}
\usage{
fitter_nn(x, params)
}
\arguments{
\item{x}{passed from fit_regressions. Defines model architecture}
}
\description{
Wrapper to Neural Network training. Defined separetely to avoid passing too many objects in parLapplyLB
}
