% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_regression_all.R
\name{fitter_svm}
\alias{fitter_svm}
\title{Wrapper to SVM training. Defined separetely to avoid passing too many objects in parLapplyLB}
\usage{
fitter_svm(x = NULL, params = NULL)
}
\arguments{
\item{x}{passed from fit_regressions}
}
\description{
Wrapper to SVM training. Defined separetely to avoid passing too many objects in parLapplyLB
}
