% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_regression_all.R
\name{fitter_xgboost}
\alias{fitter_xgboost}
\title{Wrapper to XGBoost training. Defined separetely to avoid passing too many objects in parLapplyLB}
\usage{
fitter_xgboost(x = NULL, params = NULL)
}
\arguments{
\item{x}{passed from fit_regressions}
}
\description{
Wrapper to XGBoost training. Defined separetely to avoid passing too many objects in parLapplyLB
}
