% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{generate_description}
\alias{generate_description}
\title{Generates description for a flowFrame}
\usage{
generate_description(ff)
}
\arguments{
\item{ff}{A flowframe}
}
\value{
A flowframe with description consistent with pData(ff@parameters)
}
\description{
Generates description for a flowFrame
}
