% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_logicle_transforms.R
\name{logicle_transform_input}
\alias{logicle_transform_input}
\title{Transform FCS files}
\usage{
logicle_transform_input(yvar, paths)
}
\arguments{
\item{yvar}{name of the exploratory measurement}

\item{paths}{Character vector of paths to store intput, intermediary results, outputs...}
}
\description{
Backbone measurements use a common transformation across the whole dataset. Exploratory measurements are transformed well/file-wise.
}
