% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_dimensionality_reduction.R
\name{perform_UMAP_dimensionality_reduction}
\alias{perform_UMAP_dimensionality_reduction}
\title{Wrapper to perform and export UMAP dimensionality reduction on the backbone}
\usage{
perform_UMAP_dimensionality_reduction(paths, extra_args_UMAP,
  chans = readRDS(file.path(paths["rds"], "chans.Rds")),
  preds = readRDS(file.path(paths["rds"], "svms_predictions.Rds")))
}
\arguments{
\item{paths}{Character vector of paths to store intput, intermediary results, outputs...}

\item{extra_args_UMAP}{Named list of arguments to pass to uwot:umap. Defaults to list(n_neighbors=15L,min_dist=0.2,metric="euclidean",verbose=verbose,n_epochs=1000L)}
}
\description{
Wrapper to perform and export UMAP dimensionality reduction on the backbone
}
