% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_regression_all.R
\name{predict_from_models}
\alias{predict_from_models}
\title{Predict missing measurements using fitted SVM regressions}
\usage{
predict_from_models(
  paths,
  prediction_events_downsampling,
  cores,
  chans = readRDS(file.path(paths["rds"], "chans.Rds")),
  events.code = readRDS(file.path(paths["rds"], "pe.Rds")),
  xp = readRDS(file.path(paths["rds"], "xp_transformed_scaled.Rds")),
  models = readRDS(file.path(paths["rds"], "regression_models.Rds")),
  train_set = readRDS(file.path(paths["rds"], "train_set.Rds")),
  regression_functions = NULL,
  verbose = TRUE,
  neural_networks_seed
)
}
\arguments{
\item{paths}{Character vector of paths to store intput, intermediary results, outputs...}

\item{prediction_events_downsampling}{Number of events to predict data for per file}

\item{cores}{Number of cores to use for parallel computing}

\item{chans}{vector of backbone channels' names}

\item{events.code}{vector of length nrow(xp) specifying from which well each event originates}

\item{xp}{Logicle-transformed backbone expression matrix}

\item{models}{list of list of machine learning models, created by infinityFlow:::fit_regressions}

\item{train_set}{data.frame with nrow(xp) rows, specifying which events were used to train the models. Generated by infinityFlow:::fit_regressions}

\item{regression_functions}{named list of fitter_* functions, passed from infinity_flow()}

\item{verbose}{Verbosity}

\item{neural_networks_seed}{Seed for computational reproducibility when using neural networks. Passed from infinity_flow()}
}
\description{
Predict missing measurements using fitted SVM regressions
}
