% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_svm_regression_all.R
\name{predict_svms}
\alias{predict_svms}
\title{Predict missing measurements using fitted SVM regressions}
\usage{
predict_svms(paths, prediction_events_downsampling, cores)
}
\arguments{
\item{paths}{Character vector of paths to store intput, intermediary results, outputs...}

\item{prediction_events_downsampling}{Number of events to predict data for per file}

\item{cores}{Number of cores to use for parallel computing}
}
\description{
Predict missing measurements using fitted SVM regressions
}
