% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_regression_all.R
\name{predict_wrapper}
\alias{predict_wrapper}
\title{Wrapper to predict. Defined separetely to avoid passing too many objects in parLapplyLB}
\usage{
predict_wrapper(x)
}
\arguments{
\item{x}{passed from predict_from_models}
}
\description{
Wrapper to predict. Defined separetely to avoid passing too many objects in parLapplyLB
}
