% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{split_matrix}
\alias{split_matrix}
\title{Split a matrix into a list of chunks. Faster than using split on a data.frame}
\usage{
split_matrix(mat, vector, byrow = TRUE)
}
\arguments{
\item{mat}{A matrix}

\item{vector}{A vector of length nrow(mat) if byrow=TRUE, ncol(mat) if byrow=FALSE}

\item{byrow}{if TRUE split rows, if FALSE split columns}
}
\description{
Split a matrix into a list of chunks. Faster than using split on a data.frame
}
