% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_fcs_scaling.R
\name{standardize_backbone_data_across_wells}
\alias{standardize_backbone_data_across_wells}
\title{Scaling of the backbone measurements}
\usage{
standardize_backbone_data_across_wells(yvar, paths,
  scale_function = function(x) {     scale(x, center = TRUE, scale =
  TRUE) }, xp = readRDS(file.path(paths["rds"], "xp_transformed.Rds")),
  chans = readRDS(file.path(paths["rds"], "chans.Rds")),
  events.code = readRDS(file.path(paths["rds"], "pe.Rds")))
}
\arguments{
\item{yvar}{name of the exploratory measurement}

\item{paths}{Character vector of paths to store intput, intermediary results, outputs...}

\item{scale_function}{Scaling function to apply. Should apply to a matrix (events x backbone measurements) and return a matrix of similar size. Defaults to scale(,center=TRUE,scale=FALSE)}
}
\description{
Scaling of the backbone measurements
}
