% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{data}
\name{steady_state_lung}
\alias{steady_state_lung}
\title{Subset of a massively parallel cytometry experiment of mouse lung single cells}
\format{
a flowSet containing 10 flowFrames (thus corresponding to 10 FCS files)
}
\source{
\url{https://flowrepository.org/id/FR-FCM-Z2LP}
}
\usage{
steady_state_lung
}
\description{
Subset of a massively parallel cytometry experiment of mouse lung single cells
}
\keyword{datasets}
