% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{data}
\name{steady_state_lung_annotation}
\alias{steady_state_lung_annotation}
\title{Target and isotypes annotation for the data object infinityFlow::steady_state_lnug}
\format{
a data.frame specifying the Infinity antibody targets and isotypes for each flowFrame of the steady_state_lung flowSet
}
\usage{
steady_state_lung_annotation
}
\description{
Target and isotypes annotation for the data object infinityFlow::steady_state_lnug
}
\keyword{datasets}
