% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_infinity_flow_fcs_parsing.R
\name{subsample_data}
\alias{subsample_data}
\title{Parsing FCS files}
\usage{
subsample_data(input_events_downsampling, paths, extra_args_read_FCS,
  name_of_PE_parameter)
}
\arguments{
\item{input_events_downsampling}{See \code{\link{infinity_flow}}}

\item{paths}{Character vector of paths to store intput, intermediary results, outputs...}

\item{extra_args_read_FCS}{passed to flowCore:read.FCS}

\item{name_of_PE_parameter}{Name of the exploratory measurement}
}
\description{
This function reads the input FCS files, downsample to a user-defined number of events if necessary, harmonize the input data names, save a concatenated expression matrix and the corresponding vector mapping events to their file of origin.
}
