% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_svm_regression_all.R
\name{svm_fitter}
\alias{svm_fitter}
\title{Wrapper to SVM training. Defined separetely to avoid passing too many objects in parLapply}
\usage{
svm_fitter(x)
}
\arguments{
\item{x}{passed from fit_svms}
}
\description{
Wrapper to SVM training. Defined separetely to avoid passing too many objects in parLapply
}
