% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_svm_regression_all.R
\name{svm_predictor}
\alias{svm_predictor}
\title{Wrapper to SVM predict. Defined separetely to avoid passing too many objects in parLapplyLB}
\usage{
svm_predictor(x)
}
\arguments{
\item{x}{passed from fit_svms}
}
\description{
Wrapper to SVM predict. Defined separetely to avoid passing too many objects in parLapplyLB
}
