% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_regression_all.R
\name{test_dependency}
\alias{test_dependency}
\title{Make sure a package is installed and produces an error otherwise}
\usage{
test_dependency(packageName)
}
\description{
Make sure a package is installed and produces an error otherwise
}
