% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_converting_to_csv_and_FCS.R
\name{export_data}
\alias{export_data}
\title{Exporting results}
\usage{
export_data(paths, FCS_export, CSV_export,
  chans = readRDS(file.path(paths["rds"], "chans.Rds")),
  transforms_chan = readRDS(file.path(paths["rds"],
  "transforms_chan.Rds")),
  transforms_pe = readRDS(file.path(paths["rds"], "transforms_pe.Rds")),
  xp = readRDS(file.path(paths["rds"], "xp.Rds")),
  umap = readRDS(file.path(paths["rds"], "umap.Rds")),
  events.code = readRDS(file.path(paths["rds"], "pe.Rds")),
  preds = readRDS(file.path(paths["rds"], "svms_predictions.Rds")),
  sampling = readRDS(file.path(paths["rds"], "sampling_preds.Rds")),
  a = read.csv(paths["annotation"], sep = ",", header = TRUE,
  stringsAsFactors = FALSE))
}
\arguments{
\item{paths}{Character vector of paths to store intput, intermediary results, outputs...}

\item{FCS_export}{if FALSE, no FCS export. if "concatenated", export one FCS file with all the data. if "split", export the data in the result folder under the subfolder FCS, with each file corresponding to a (subsampled) input file.}
}
\description{
Exporting results
}
