% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_regression_all.R
\name{fit_regressions}
\alias{fit_regressions}
\title{Train SVM regressions}
\usage{
fit_regressions(
  yvar,
  params,
  paths,
  cores,
  xp = readRDS(file.path(paths["rds"], "xp_transformed_scaled.Rds")),
  chans = readRDS(file.path(paths["rds"], "chans.Rds")),
  events.code = readRDS(file.path(paths["rds"], "pe.Rds")),
  transforms_chan = readRDS(file.path(paths["rds"], "transforms_chan.Rds")),
  transforms_pe = readRDS(file.path(paths["rds"], "transforms_pe.Rds")),
  regression_functions,
  verbose = TRUE,
  neural_networks_seed
)
}
\arguments{
\item{yvar}{name of the exploratory measurement}

\item{params}{Named list of arguments passed to regression fitting functions}

\item{paths}{Character vector of paths to store intput, intermediary results, outputs...}

\item{cores}{Number of cores to use for parallel computing}

\item{xp}{Logicle-transformed backbone expression matrix}

\item{chans}{vector of backbone channels' names}

\item{events.code}{vector of length nrow(xp) specifying from which well each event originates}

\item{transforms_chan}{named list of logicle-transformations for backbone channels}

\item{transforms_pe}{named list of logicle-transformations for Infinity channels}

\item{regression_functions}{named list of fitter_* functions, passed from infinity_flow()}

\item{verbose}{Verbosity}

\item{neural_networks_seed}{Seed for computational reproducibility when using neural networks. Passed from infinity_flow()}
}
\description{
Train SVM regressions
}
