% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_regression_all.R
\name{fitter_glmnet}
\alias{fitter_glmnet}
\title{Wrapper to glmnet. Defined separetely to avoid passing too many objects in parLapplyLB}
\usage{
fitter_glmnet(x, params)
}
\arguments{
\item{x}{passed from fit_regressions}
}
\description{
Wrapper to glmnet. Defined separetely to avoid passing too many objects in parLapplyLB
}
