% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_regression_all.R
\name{predict_from_models}
\alias{predict_from_models}
\title{Predict missing measurements using fitted SVM regressions}
\usage{
predict_from_models(paths, prediction_events_downsampling, cores,
  chans = readRDS(file.path(paths["rds"], "chans.Rds")),
  events.code = readRDS(file.path(paths["rds"], "pe.Rds")),
  models = readRDS(file.path(paths["rds"], "regression_models.Rds")),
  xp = readRDS(file.path(paths["rds"], "xp_transformed_scaled.Rds")),
  train_set = readRDS(file.path(paths["rds"], "train_set.Rds")))
}
\arguments{
\item{paths}{Character vector of paths to store intput, intermediary results, outputs...}

\item{prediction_events_downsampling}{Number of events to predict data for per file}

\item{cores}{Number of cores to use for parallel computing}
}
\description{
Predict missing measurements using fitted SVM regressions
}
